Make com.devture.ansible.role.postgres optional
Until now, we would call into it to determine the Postgres version.

With these changes, it's no longer required.

Like `defaults/main.yml` was saying in a comment, we now default to
`:latest` if the Postgres version is not set.

If we get pointed to the devture-postgres role via `backup_borg_postgresql_version_detection_devture_postgres_role_name`,
we'll use it. Otherwise, we complain that we can't auto-detect the
version and that it's not set manually either.